fix(dash): Waveshare Dashboard Patches v2
- Remove Tachometer dependency and compute metrics using total_distance.

- Fix negative speed bug by applying fabs() to parsed ERPM and speed calculation.

- Prevent string-level jitter on UI labels using buffered snprintf and strcmp.

- Upgrade total_distance to double for precise mileage tracking.

+++ include/odometer.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // Global odometer value (in km)
-extern float total_distance;
+extern double total_distance;
 
 class Odometer {
 public:
